Have singleton org.embulk.EmbulkDependencyClassLoader merged from EmbulkDependencyClassLoaders and DependencyClassLoader
