+++ embulk-core/build.gradle
@@ -65,7 +65,7 @@ dependencies {
     testImplementation project(":embulk-junit4")
     testImplementation "ch.qos.logback:logback-classic:1.3.6"
 
-    // TODO: Remove this, and load it with the proper DependencyClassLoader.
+    // TODO: Remove this, and load it with EmbulkDependencyClassLoader.
     // This statement gets it loaded by the top-level ClassLoader.
     testImplementation project(":embulk-deps")
 }
