+++ embulk-core/src/main/java/org/embulk/cli/CommandLineParser.java
@@ -3,7 +3,7 @@ package org.embulk.cli;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import org.embulk.deps.EmbulkDependencyClassLoaders;
+import org.embulk.EmbulkDependencyClassLoader;
 import org.slf4j.Logger;
 
 public abstract class CommandLineParser {
@@ -35,7 +35,7 @@ public abstract class CommandLineParser {
         }
     }
 
-    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoaders.get();
+    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoader.get();
     private static final String CLASS_NAME = "org.embulk.deps.cli.CommandLineParserImpl";
 
     static {
