+++ embulk-core/src/main/java/org/embulk/config/ConfigLoaderDelegate.java
@@ -7,7 +7,7 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Map;
 import java.util.Properties;
-import org.embulk.deps.EmbulkDependencyClassLoaders;
+import org.embulk.EmbulkDependencyClassLoader;
 
 public abstract class ConfigLoaderDelegate {
     public static ConfigLoaderDelegate of(final ModelManagerDelegate model) {
@@ -56,7 +56,7 @@ public abstract class ConfigLoaderDelegate {
         }
     }
 
-    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoaders.get();
+    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoader.get();
     private static final String CLASS_NAME = "org.embulk.deps.config.ConfigLoaderDelegateImpl";
 
     static {
