+++ embulk-core/src/main/java/org/embulk/config/ModelManagerDelegate.java
@@ -2,7 +2,7 @@ package org.embulk.config;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
-import org.embulk.deps.EmbulkDependencyClassLoaders;
+import org.embulk.EmbulkDependencyClassLoader;
 
 public abstract class ModelManagerDelegate {
     public static ModelManagerDelegate of() {
@@ -49,7 +49,7 @@ public abstract class ModelManagerDelegate {
         }
     }
 
-    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoaders.get();
+    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoader.get();
     private static final String CLASS_NAME = "org.embulk.deps.config.ModelManagerDelegateImpl";
 
     static {
