+++ embulk-core/src/main/java/org/embulk/plugin/maven/MavenArtifactFinder.java
@@ -4,7 +4,7 @@ import java.io.FileNotFoundException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Path;
-import org.embulk.deps.EmbulkDependencyClassLoaders;
+import org.embulk.EmbulkDependencyClassLoader;
 import org.embulk.plugin.MavenPluginType;
 
 public abstract class MavenArtifactFinder {
@@ -46,7 +46,7 @@ public abstract class MavenArtifactFinder {
         }
     }
 
-    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoaders.get();
+    private static final ClassLoader CLASS_LOADER = EmbulkDependencyClassLoader.get();
     private static final String CLASS_NAME = "org.embulk.deps.maven.MavenArtifactFinderImpl";
 
     static {
